+++ client/src/App.js
@@ -1,14 +1,18 @@
 import React, { Component } from 'react';
 import Routes from './Routes';
 import 'bootstrap/dist/css/bootstrap.css';
+import { Provider } from "react-redux";
 
+import store from "./store"
 
 class App extends Component {
   render() {
     return (
-      <div style={{backgroundColor: "#191919", minHeight: "100vh"}}>
-        <Routes />
-      </div>
+      <Provider store={store}>
+        <div style={{ backgroundColor: "#191919", minHeight: "100vh" }}>
+          <Routes />
+        </div>
+      </Provider>
     )
   }
 }