+++ client/src/actions/types.js
@@ -1,0 +1,6 @@
+export const GET_911_CALLS = 'GET_911_CALLS';
+
+// BOLOS
+export const GET_BOLOS = 'GET_BOLOS';
+export const CREATE_BOLO = 'CREATE_BOLO';
+export const REMOVE_BOLO = 'REMOVE_BOLO';