+++ client/src/reducers/index.js
@@ -1,0 +1,7 @@
+/* disable-eslint */
+import { combineReducers } from "redux";
+import boloReducer from "./boloReducer";
+
+export default combineReducers({
+    bolos: boloReducer
+});